+++ adlfs/spec.py
@@ -3,8 +3,9 @@
 
 from __future__ import absolute_import, division, print_function
 
-import io
+import asyncio
 from glob import has_magic
+import io
 import logging
 import os
 import warnings
@@ -26,7 +27,7 @@ from fsspec.asyn import (
 )
 from fsspec.spec import AbstractBufferedFile
 from fsspec.utils import infer_storage_options, tokenize
-from .utils import filter_blobs, get_blob_metadata, close_container_client, close_service_client
+from .utils import filter_blobs, get_blob_metadata, close_client
 
 
 logger = logging.getLogger(__name__)
@@ -401,7 +402,7 @@ class AzureBlobFileSystem(AsyncFileSystem):
         else:
             self.sync_credential = None
         self.do_connect()
-        weakref.finalize(self, maybe_sync, close_service_client, self, self)
+        weakref.finalize(self, maybe_sync, close_client, self, self)
 
     @classmethod
     def _strip_protocol(cls, path: str):
@@ -1646,7 +1647,9 @@ class AzureBlobFile(AbstractBufferedFile):
 
     def close(self):
         """Close file and azure client."""
-        maybe_sync(close_container_client, self, self)
+        # asyncio.run_coroutine_threadsafe(self.container_client.close(), loop=self.loop)
+        asyncio.run_coroutine_threadsafe(close_client(self), loop=self.loop)
+
         super().close()
 
     def connect_client(self):