+++ adlfs/tests/test_spec.py
@@ -69,7 +69,7 @@ def assert_blobs_equals(blobs, expected_blobs):
 
 def test_ls(storage):
     fs = AzureBlobFileSystem(
-        account_name=storage.account_name, connection_string=CONN_STR
+        account_name=storage.account_name, connection_string=CONN_STR,
     )
 
     ## these are containers
@@ -566,7 +566,7 @@ def test_open_file(storage, mocker):
 
     future = asyncio.Future()
     future.set_result("close")
-    close = mocker.patch.object(f.container_client, "close", return_value=future)
+    close = mocker.patch.object(f.container_client, "close", return_value="close")
     f.close()
     print(fs.ls("/data/root/a"))
 
@@ -581,7 +581,7 @@ def test_open_context_manager(storage, mocker):
     future = asyncio.Future()
     future.set_result("close")
     with fs.open("/data/root/a/file.txt") as f:
-        close = mocker.patch.object(f.container_client, "close", return_value=future)
+        close = mocker.patch.object(f.container_client, "close", return_value="close")
         result = f.read()
         assert result == b"0123456789"